Split merge command
